Changed logic of the product selection to show a single granule at a time
Major restructurization of the app. Removed layers module, as it is
not needed anymore.
The logic is now focused in module map with data provided by module
products.

User can only select a single granule per product type. Selecting
a granule of an exisiting type again results in modyfing the granule
view only.