Change API prefix from 'api' to 'api/v1'
While some SO answers argue, that keeping the version in the api path is
against the REST principles, I think it is worth to disregard this.
In our case, the API won't really be fully REST compliant and in the
case we have to roll out new API version, we will be able to transition
more easily.

Fixes: #24