[#93] Use InputStream to read SLD file
Previously, the code assumed, that the read file is placed somewhere in
the fs, and not in a JAR file as a resource, while the latter is the
case.

This caused the method to fail when the app was run from a JAR file,
now I use an InputStream, which works in both cases.