[#46] Add Places search endpoint `/api/v1/places`
The Place entity maps useful fields of the PRNG miejscowosci.xlsx
source.
The original xlsx is a binary file, which shouldn't go into the git repo.
To fix this, I created a tool PlacesXlsxToCsv, which transforms it
into a CSV file, which is still quite large but is a text file, which
contains only the fields from the entity.

The CSV file is committed into the repo.
In case it has to be updated, it can be regenerated from an updated
miejscowosci.xlsx source using the tool.

The CSV file is used to populate the DB in the seeding phase.
To speed up seeding two threads are allowed in ThreadConfig.
Also, as the threads from the config are only used for seeding phase for
now, I allow them to be terminated after timeout
(setAllowCoreThreadTimeOut=true in ThreadConfig).

I created an endpoint `/api/v1/places` which returns paginated
results for passed name prefix.
I limit the maximum page size to 100 entries in `application.properties`.