[#103] Serve ProductType and Overlay icons
Serve the images using minio/ceph.
The icons are put in the repo with names matching ProductTypes and
Overlays.

The README.md is updated to reflect the changes.
Switch from s3cmd to mc, as the latter supports easier policy management.

Add unique constraint on names of ProductTypes and Overlays to avoid
problems in case of accidentally creating duplicates.