[#29] Add user registration
I added AppUser.enabled field.
It is now by default set to true, but once email confirmation is set up
it should be set to false, and set to true after the email is confirmed.

Also, AppUser.email is given a unique constraint to prevent duplicate
entries.
The JPA exception is mapped to AppUserCreationException in AppUserService.
ErrorHandler returns status 200 in such case to prevent email
enumeration attack.

I created AppUserController, which handles registration under `/register`
endpoint.

Register- and LoginRequest get validation constraints.
Password for now only has length constraint.

The validation errors are handled in ErrorHandler and mapped to a useful
form using ErrorHandlerHelper.
The errors are translated to Polish using ResourceBundleMessageSource
and a file with translations.

AppRoleSetConverted has added tests, as it turned out it failed if
an AppUser had no roles.