[#28] Email verification
Add token based email verification. An AppUser is created with field
enabled=false, and it can be activated using the token sent by mail.

The logic of sending emails is decoupled from the controller to async
event listeners.

Create LoggingMailSender, which outputs email contents to log.

For testing I introduced GreenMail, which is used to verify appropriate
emails are sent.

The email contents are put in messages_pl_PL.properties.