[#120] reCAPTCHA backend
Add dependency to recaptcha-spring-boot-starter, which handles
sending request to the Google reCAPTCHA service.

The RecaptchaValidator is called in the AppUserController::register,
validating the received param `g-recaptcha-response`.

In case of validator error appropriate error codes are returned in
field `recaptcha`.

For tests the validator test mode is enabled, and in development it
uses the always-returning-true key-set which is recommended for tests in
reCAPTCHA FAQ.

Add field `recaptchaSiteKey` to the `/api/v1/config` response.