[#307] Support timezone parameter in scenes endpoints
Add `tz` parameter to both SceneController endpoints, which
handling of day boundaries change depending on it.
The date-times in responses are in line with the specified `tz`.

Its values are keys of the tz database entries,
https://www.iana.org/time-zones.
The java runtime uses the latest available version of the db at the time
of installation (see https://stackoverflow.com/a/42513531).

The SceneRepository::findDatesWithData is removed, in favor of a less
efficient but timezone-aware solution in
SceneService::getAvailabilityDates.

Prune StringToYearMonthConverter, which was unused.