[#371] Modify Product and Scene to match with GeoServer ImageMosaic
Move the layer_name from Scene to Product, as we will have a single
layer per Product after this change.
Drop Scene.created, as it won't apply anymore.
Add granule_path and footprint to Scene, which are required by IM.

There is a lot of fixes in tests, as the newly added fields are required
to be non-null.

The WebhookController no longer communicates directly with GS.

The SeedProducts now only creates views named `scene_<product_name>`.
The stores and layers are left not-configured for now, the
implementation should be put in GeoServerService::addLayer.
A scene is not created if the corresponding GeoTiff doesn't exist in S3,
the process will fail if the S3 connection is misconfigured.