[#371] Update GeoServer to 1.2.0-GS2.16.2 and correct GS endpoint name
The updated GeoServer uses a patched GeoTools version, so that
ImageMosaic+S3Geotiff works.

The endpoint is set to `mailto`, so that an s3geotiff location is a
correct URL, e.g. `mailto://bucket/key`. The protocol part here only
differentiates endpoint configurations, so it doesn't matter if the rest
isn't a correct email address.