[#579] Admin Product controller
Add endpoints to manage Products.
For now, cases such as changing granuleRuleArtifact don't propagate to
Scenes, assuming there will be another mechanism to trigger such
re-synchronization.

The column product.searchable_metadata was conceived to hold info about
which metadata is searchable, but we went in a different direction,
hence the drop migration.