[#621] Add admin Scenes synchronization endpoints
The synchronization system is divided into three main parts:
- SyncJobManager: manages jobs state, running operation and details,
- ChunkedRunner: processes a stream of Runnables in chunks,
- SyncJobStore: maintains a list of jobs.

The idea behind using ChunkedRunner is not to load all the scenes keys
into memory, as they can be loaded in chunks while maintaining
parallelism.

During the operation of a job it maintains its state, which is exposed
under GET endpoint - in this way it's possible to monitor the operation
of the job.
For example, it could fail for all keys - in that case it can be
cancelled, and the problem investigated.
All the acceptor Errors are stored in the job for inspection.

The jobs are not persisted to DB, and are the state of a specific
instance of backend.
Memory can be reclaimed by deleting jobs.