[#623] Improve DB normalization
Remove s3path and granule_path from table scene, moving their
calculation to a dedicated view: scene_extended.
This view, on the other hand, is used to create views for GeoServer.

A granule_path consists of a prefix and a suffix.
The prefix is now stored as a row in table property, and is initialized
in SeedProducts.
In case no seed were used, admin can modify it's value through endpoints.

The suffix is on the other hand a product of metadata_content,
scene_content and granule_artifact_rule.
It's calculation is moved to a db function, and results go to the
intermediate view scene_extended.

Moreover, timestamp is now a generated field, which is possible as it
only depends on metadata_content.
It is still independently indexed and kept as a column, to allow easier
repository usage.

There is also a cascade of changes in other places, as now scene_content
and metadata_content must always be set.