[#665] Add Products seed script
The script replaces SeedProducts, integrating in a different lifecycle
moment of the application - it is meant to be run independently of
running specific backend instance.

The seed script operation is configurable from a seed-config: one is
placed in s4e-backend/seed-configs/seed-1.
It is also used in e2e tests, instead of the previous s4e-test-1
dataset, especially when SeedProducts is decommissioned.
Inclusion in e2e tests also ensures it works.

Change in the e2e tests also required few corrections, so that it works
with updated date (now, data in e2e test is from 2020-02-01).

Moreover, configure the default proxy configuration to pass directly to
/api/v1, requiring direct backend access for /api/admin prefix.

Create helper script copy-scenes.sh, which makes it easier to prepare a
bogus data package.
It only copies metadata and scene files, while touching all other
artifacts.