[#288] Introduce LicenseGrants to handle private Products
Add Product.accessType, which determines the type of license it has.

Add entity LicenseGrant, which is to be used with private license.
It is relevant for Products with accessType PRIVATE.

Allow to set the Product's license grant from admin endpoints, making
sure to clean up lingering grants if the access is changed not to
PRIVATE.

Introduce the accessType field in seed-1.
For now, all products are OPEN: this default for existing Products is
also implemented in the migration.