[#288] Add license handling
LicensePermissionEvaluator is central to the license handling: it
has an interface for evaluating access to Products and Scenes.
These checks are done based on id and authentication, and they are
utilized throughout the HTTP security (SecurityConfig), ProductService
and SceneService.

The implementation of the actual rules for checking the access is put in
a separate class per license type, i.e., OpenLicense, EumetsatLicense
and PrivateLicense.

Checking for permissions requires some overhead, so when possible it's
avoided.
Only EUMETSAT license requires checking each scene, as OPEN doesn't
introduce any restriction and PRIVATE does the check on the Product
level - also, ROLE_ADMIN is assumed to always have access.
Nevertheless, few methods require checking scenes one by one.

Search endpoints are not secured, but their corresponding download
endpoints are so that it's not possible to download the data without
license.

Add licensed layers info to the JWT token, so that GeoServer gateway can
selectively allow access based on it.
Moreover, expose a claim which can be used to allow privileged access
(for ZK and such).