[#706] Move expert email configuration to a DB property
After consultation, the expert email shouldn't be defined in a runtime
property, but rather in a DB property.
Change the mechanism for it's storage.

In case of a missing property an error is logged when trying to send the
request.

Improve mail generation a little, setting a proper Reply-To in the email
sent to the expert.