[#862] Add ZK option to Institution
Add a boolean field zk to Institution and propagate it in the hierarchy.

The field should be settable by admin in AdminInstitutionController.
Allow only to unset it on the root of ZK (that is on an institution
which has no parent or its parent has zk==false).

Use it to populate ROLE_ZK_MEMBER.
Remove obsolete AppUser.zkMember.