[#1022] Add polygon to SceneResponse
Move mapping operations to a separate mapper.
Add polygon to SceneResponse, its mapping and a test for correct
operation.

Promote a footprint mapping exception to WARN log call in
AdminSceneMapper, as it's an erroneous behaviour which should be caught.
Use the same log level in new SceneMapper footprint mapping.

Correct the field order in SceneResponse to match with the Scene bean.