[#716] Remove sentry.properties and correct log timestamp pattern
The sentry.properties files are not necessary, and only generate
stacktraces in runtime. If removed, Sentry issues a log entry at WARN
level anyway.

Set a logback timestamp pattern which complies with ISO8601, see e.g. SO
https://stackoverflow.com/a/23096247/1778478 for more info (beware
though, as there's an error in this SO answer - the fraction of a
second should be dot (.) and not comma (,) delimited).