[#1039] Search based on Product.name instead of metadata.product_type
Switch to using Product.name by modifying the issued query.
Move handling of the parameter from QueryText to a new QueryProductType
class.

Drop the index on metadata_content.product_type, as it's no longer used
in search queries.

Pre-build the static select and count queries in static initializer of
QueryBuilderImpl.

Implement various IDE suggested code improvements.
Consolidate query creation, so that all appended portions start with a
space, in this way there's no need to leave trailing spaces in some
places.

Use f_cast_isots instead of directly using to_timestamp, removing the
duplication of the date format and using the same function which on
which index is created.