[#1111] Make LoggingMailSender registration configurable by a property
The property is named: `mail.logging-mail-sender.enable`, and by default
it is set to false (only in profile development, true).

The LogginMailSender may be undesirable in development profile when
running e2e tests (and you want to connect to an actual SMTP server).
Switching to profile production in e2e tests leads to timeouts in some
difficult to correct places and it may be a topic for further
optimization.